scala client conform to  apache style
