+++ scala-client/README.md
@@ -1,8 +1,24 @@
+<!--
+Licensed to the Apache Software Foundation (ASF) under one
+or more contributor license agreements.  See the NOTICE file
+distributed with this work for additional information
+regarding copyright ownership.  The ASF licenses this file
+to you under the Apache License, Version 2.0 (the
+"License"); you may not use this file except in compliance
+with the License.  You may obtain a copy of the License at
+
+http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing,
+software distributed under the License is distributed on an
+"AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+KIND, either express or implied.  See the License for the
+specific language governing permissions and limitations
+under the License.
+-->
 # pegasus-scala-client
 
-This is the scala client for [xiaomi/pegasus](https://github.com/XiaoMi/pegasus).
-
-It's built on top of the java client [xiaomi/pegasus-java-client](https://github.com/XiaoMi/pegasus-java-client).
+It's built on top of the java client [apache/incubator-pegasus/java-client](https://github.com/apache/incubator-pegasus/tree/master/java-client).
 
 ## Features:
 
@@ -13,7 +29,7 @@ It's built on top of the java client [xiaomi/pegasus-java-client](https://github
 ## Example:
 
 ```scala
-    import com.xiaomi.infra.pegasus.scalaclient.Serializers._
+    import org.apache.pegasus.scalaclient.Serializers._
 
     val c = ScalaPegasusClientFactory.createClient("resource:///pegasus.properties")
