+++ scala-client/scripts/ci-test.sh
@@ -1,5 +1,24 @@
 #!/usr/bin/env bash
 
+#
+# Licensed to the Apache Software Foundation (ASF) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The ASF licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied.  See the License for the
+# specific language governing permissions and limitations
+# under the License.
+#
+
 set -e
 
 SCRIPT_DIR=$(dirname "${BASH_SOURCE[0]}")
@@ -21,9 +40,9 @@ shellcheck_must_pass ./scripts/ci-test.sh
 sbt scalafmtSbtCheck scalafmtCheck test:scalafmtCheck
 
 # install java-client dependency
-git clone https://github.com/XiaoMi/pegasus-java-client.git
-cd pegasus-java-client
-git checkout v2.2.0
+git clone https://github.com/apache/incubator-pegasus.git
+cd incubator-pegasus/java-client
+git checkout master
 mvn clean package -DskipTests -Dcheckstyle.skip=true
 mvn clean install -DskipTests -Dcheckstyle.skip=true
 cd ..
