+++ scala-client/src/main/scala/org/apache/pegasus/scalaclient/Models.scala
@@ -1,9 +1,28 @@
-package com.xiaomi.infra.pegasus.scalaclient
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
 
-import com.xiaomi.infra.pegasus.client.PException
-import com.xiaomi.infra.pegasus.client.{FilterType, PException}
-import com.xiaomi.infra.pegasus.scalaclient.{Serializer => SER}
+package org.apache.pegasus.scalaclient
+
+import org.apache.pegasus.client.PException
+import org.apache.pegasus.client.{FilterType, PException}
 import scala.concurrent.duration._
+import org.apache.pegasus.scalaclient.{Serializer => SER}
 
 /**
   * [Copyright]
