+++ scala-client/src/main/scala/org/apache/pegasus/scalaclient/ScalaPegasusClient.scala
@@ -1,12 +1,31 @@
-package com.xiaomi.infra.pegasus.scalaclient
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package org.apache.pegasus.scalaclient
 
 import java.util.Properties
-import com.xiaomi.infra.pegasus.client.{
+import org.apache.pegasus.client.{
   PException,
   TableOptions,
   PegasusClientInterface => IClient
 }
-import com.xiaomi.infra.pegasus.scalaclient.{Serializer => SER}
+import org.apache.pegasus.scalaclient.{Serializer => SER}
 
 import scala.concurrent.duration._
