+++ scala-client/src/test/resources/log4j.properties
@@ -1,18 +1,33 @@
+#
+# Licensed to the Apache Software Foundation (ASF) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The ASF licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied.  See the License for the
+# specific language governing permissions and limitations
+# under the License.
+#
 log4j.rootLogger=DEBUG, rolling_file
-log4j.logger.com.xiaomi.infra.pegasus = INFO
-
+log4j.logger.org.apache.pegasus=INFO
 log4j.appender.stdout=org.apache.log4j.ConsoleAppender
 log4j.appender.stdout.Target=System.out
 log4j.appender.stdout.layout=org.apache.log4j.PatternLayout
 log4j.appender.stdout.layout.ConversionPattern=%d{yyyy-MM-dd HH:mm:ss} %-5p %c{1}:%L - %m%n
-
 log4j.appender.file=org.apache.log4j.FileAppender
 log4j.appender.file.File=log.txt
 log4j.appender.file.ImmediateFlush=true
 log4j.appender.file.layout=org.apache.log4j.PatternLayout
 log4j.appender.file.layout.ConversionPattern=%d{yyyy-MM-dd HH:mm:ss} %-5p %c{1}:%L - %m%n
 log4j.appender.file.Append=true
-
 log4j.appender.rolling_file=org.apache.log4j.RollingFileAppender
 log4j.appender.rolling_file.File=rolling_log/log.txt
 log4j.appender.rolling_file.MaxFileSize=10MB
