+++ scala-client/src/test/resources/pegasus.properties
@@ -1,6 +1,24 @@
-meta_servers = 127.0.0.1:34601,127.0.0.1:34602,127.0.0.1:34603 
-operation_timeout = 1000
-async_workers = 4
-enable_perf_counter = false
-perf_counter_tags = cluster=onebox,app=unit_test
-push_counter_interval_secs = 10
+#
+# Licensed to the Apache Software Foundation (ASF) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The ASF licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied.  See the License for the
+# specific language governing permissions and limitations
+# under the License.
+#
+meta_servers=127.0.0.1:34601,127.0.0.1:34602,127.0.0.1:34603 
+operation_timeout=1000
+async_workers=4
+enable_perf_counter=false
+perf_counter_tags=cluster=onebox,app=unit_test
+push_counter_interval_secs=10
